This is the 6.12.40 stable release

Resolved conflicts in:
drivers/crypto/xilinx/zynqmp-sha.c
drivers/firmware/xilinx/zynqmp-pm.c
drivers/gpu/drm/drm_fbdev_dma.c
drivers/i2c/i2c-atr.c
drivers/net/ethernet/xilinx/xilinx_axienet_main.c
drivers/pci/controller/pcie-xilinx-cpm.c
drivers/spi/spi-zynqmp-gqspi.c
drivers/usb/dwc3/core.h
drivers/usb/dwc3/gadget.c
drivers/usb/host/xhci.c
drivers/usb/misc/onboard_usb_dev.c

+++ arch/riscv/include/asm/cmpxchg.h
@@ -169,7 +169,7 @@
 		break;							\
 	case 4:								\
 		__arch_cmpxchg(".w", ".w" sc_sfx, prepend, append,	\
-				__ret, __ptr, (long), __old, __new);	\
+				__ret, __ptr, (long)(int)(long), __old, __new);	\
 		break;							\
 	case 8:								\
 		__arch_cmpxchg(".d", ".d" sc_sfx, prepend, append,	\
